Cheap Dimmer Switch.
A dimmer switch controls the brightness of one or more bulbs.
You have a cheap dimmer switch, that its level goes from 5 to 15.

A bulb has a dynamic wattage. Its brightness is based on the wattage
and the level of the switch.

0 - 10
5 - 15

MIN = 5
MAX = 15

5        10    12    15
|-------------------|

switch - bulb
8      - bulb
       - bulb

| Dimmer switch level | Bulb with 30 wattage | Bulb with 50 wattage | Bulb with 70 wattage |
| ------------------- | -------------------- | -------------------- | -------------------- |
| 5                   | 0 brightness         | 0 brightness         | 0 brightness         |
| 10                  | 15 brightness        | 25 brightness        | 35 brightness        |
| 12                  | ?  br                | ?
| 15                  | 30 brightness        | 50 brightness        | 70 brightness        |

Implement these entities so the test cases below pass.